rework connect related stuff
handle our own connection states, so we don't have to rely on the crappy
BluetoothManager.getConnectionState().
the problem is, if we try BluetoothManager.getConnectionState(this,
BluetoothProfile.GATT), it returns true if there're other on-going
connections, even from other app...
but we only cares about our own.

so, track the connection state by ourselfs...